% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorblind_friendly.R
\name{colorblind_friendly}
\alias{colorblind_friendly}
\title{Returns a data frame of colorblind friendly colour palettes}
\usage{
colorblind_friendly(palettes = PrettyColsPalettes)
}
\arguments{
\item{palettes}{List of palettes. Default PrettyColsPalettes.}
}
\value{
A data frame with two columns giving names and types of colourblind
friendly palettes.
}
\description{
Returns a data frame of colorblind friendly colour palettes
}
\examples{
colorblind_friendly()
}
