% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourblind_friendly.R
\name{colourblind_friendly}
\alias{colourblind_friendly}
\title{Returns a data frame of colourblind friendly colour palettes}
\usage{
colourblind_friendly(palettes = PrettyColsPalettes)
}
\arguments{
\item{palettes}{List of palettes. Default PrettyColsPalettes.}
}
\value{
A data frame with two columns giving names and types of colourblind
friendly palettes.
}
\description{
Returns a data frame of colourblind friendly colour palettes
}
\examples{
colourblind_friendly()
}
