% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complementary_color.R
\name{complementary_color}
\alias{complementary_color}
\title{Find complementary color of a hex code}
\usage{
complementary_color(hex)
}
\arguments{
\item{hex}{Hex code given as a string}
}
\value{
c_col hex code of complementary color
}
\description{
Find complementary color of a hex code
}
