% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complementary_colour.R
\name{complementary_colour}
\alias{complementary_colour}
\title{Find complementary colour of a hex code}
\usage{
complementary_colour(hex)
}
\arguments{
\item{hex}{Hex code given as a string}
}
\value{
c_col hex code of complementary colour
}
\description{
Find complementary colour of a hex code
}
