% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrettyCols-package.R
\docType{package}
\name{PrettyCols-package}
\alias{PrettyCols}
\alias{PrettyCols-package}
\title{PrettyCols: Pretty Colour Palettes}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Defines aesthetically pleasing colour palettes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nrennie.rbind.io/PrettyCols/}
  \item \url{https://github.com/nrennie/PrettyCols}
  \item Report bugs at \url{https://github.com/nrennie/PrettyCols/issues}
}

}
\author{
\strong{Maintainer}: Nicola Rennie \email{nrennie35@gmail.com}

}
\keyword{internal}
