% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_pretty_d.R
\name{scale_colour_pretty_d}
\alias{scale_colour_pretty_d}
\title{Plotting with PrettyCols palettes for colour ggplot2}
\usage{
scale_colour_pretty_d(name, direction = 1)
}
\arguments{
\item{name}{Name of Palette. Run \code{names(PrettyColsPalettes)} to view options.}

\item{direction}{Sets order of colors. Default palette is 1. If direction is -1, palette color order is reversed}
}
\description{
Plotting with PrettyCols palettes for colour ggplot2
}
