% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettycols.R
\name{prettycols}
\alias{prettycols}
\title{Generates the colour palettes}
\usage{
prettycols(name, n, type = "discrete", direction = 1)
}
\arguments{
\item{name}{Name of Palette. Choices are:
\code{MoodyBlues}}

\item{n}{Number of desired colors. If number of requested colors is beyond the scope of the palette,
colors are automatically interpolated. If n is not provided, the length of the palette is used.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want to automatically
interpolate between colors. Default "discrete"}

\item{direction}{Sets order of colors. Default palette is 1. If direction is -1, palette color order is reversed}
}
\value{
A vector of colors.
}
\description{
Generates the colour palettes
}
\examples{
prettycols("MoodyBlues")
}
