% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrettyColsPalettes.R
\docType{data}
\name{PrettyColsPalettes}
\alias{PrettyColsPalettes}
\title{Complete list of palettes.}
\format{
An object of class \code{list} of length 22.
}
\usage{
PrettyColsPalettes
}
\description{
Use names(PrettyColsPalettes) to return all palette names.
}
\keyword{datasets}
