% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_pretty_c.R
\name{scale_color_pretty_c}
\alias{scale_color_pretty_c}
\title{Plotting with PrettyCols palettes for colour ggplot2}
\usage{
scale_color_pretty_c(name, direction = 1, ...)
}
\arguments{
\item{name}{Name of Palette. Run \code{names(PrettyColsPalettes)} to view options.}

\item{direction}{Sets order of colors. Default palette is 1. If direction is -1,
palette color order is reversed}

\item{...}{Other arguments passed on to \code{\link[ggplot2]{scale_color_gradientn}}}
}
\value{
A ggproto object defining a continuous colour scale for use with ggplot2.
}
\description{
Plotting with PrettyCols palettes for colour ggplot2
}
\examples{
library(ggplot2)
ggplot(data=iris, aes(x=Petal.Width, y=Petal.Length, color=Petal.Width)) +
  geom_point() +
  scale_color_pretty_c("Tangerines")
}
