% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_pretty_c.R
\name{scale_color_pretty_c}
\alias{scale_color_pretty_c}
\title{Plotting with PrettyCols palettes for colour ggplot2}
\usage{
scale_color_pretty_c(name, direction = 1)
}
\arguments{
\item{name}{Name of Palette. Choices are: \code{MoodyBlues}}

\item{direction}{Sets order of colors. Default palette is 1. If direction is -1, palette color order is reversed}
}
\description{
Plotting with PrettyCols palettes for colour ggplot2
}
