% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_all_palettes.R
\name{view_all_palettes}
\alias{view_all_palettes}
\title{Prints all available colour palettes}
\usage{
view_all_palettes(type = "all")
}
\arguments{
\item{type}{Type of colour palettes to view. Either all palettes, sequential palettes,
diverging palettes, or qualitative palettes. One of c("all", "seq", "div", "qual").
Default "all".}
}
\value{
A plot of all colour palettes available in the package.
}
\description{
Prints all available colour palettes
}
\examples{
view_all_palettes(type = "seq")
}
